Change useResources to use react-query
